requirements.txt
absl-py==2.3.1
attrs==25.3.0
chex==0.1.90
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.9
fonttools==4.55.8
gast==0.6.0
glcontext==3.0.0
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.8
matplotlib==3.10.0
ml_dtypes==0.5.3
moderngl==5.12.0
numpy==2.2.2
opt_einsum==3.4.0
optax==0.2.5
packaging==24.2
pillow==11.1.0
polars==1.32.3
pycapnp==2.0.0
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
rlax==0.1.7
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
tensorflow-probability==0.25.0
tfp-nightly==0.26.0.dev20250823
toolz==1.0.0
typing_extensions==4.14.1
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
